move utils to their own folder; cleanup server.js
